feat: rename project and update configurations, setup tailwind css
The project has been renamed from "Vite App" to "ChatLuna" and various configurations
have been updated to reflect this change. The title in the index.html has been updated,
as well as the project dependencies in package.json to include tailwind CSS and other
components essential for the new project direction. Additionally, the .prettierrc.json
has been adjusted to modify code styling settings in line with the new project requirements.